// src/core/config/build-config/types/types.ts
export interface BuildPaths {
    entry: string;
    html: string;
    public: string;
    output: string;
    src: string;
}
export type BuildMode = 'production' | 'development';
export type BuildPlatform = 'mobile' | 'desktop';

export interface BuildOptions {
    port: number;
    paths: BuildPaths;
    mode: BuildMode;
    platform: BuildPlatform;
    analyzer?: boolean;
}